profile, update profile, change password, email verification added

+++ app/Http/Controllers/User/UserController.php
@@ -2,13 +2,16 @@
 
 namespace App\Http\Controllers\User;
 
-use App\Http\Controllers\Controller;
-use App\Http\Requests\user\profile\UpdateRequest;
-use App\Models\Profile;
 use App\Models\User;
+use App\Models\Profile;
+use Illuminate\Support\Str;
 use Illuminate\Http\Request;
+use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\File;
+use Illuminate\Support\Facades\Hash;
+use App\Http\Requests\user\profile\ChangePassword;
+use App\Http\Requests\user\profile\UpdateProfileRequest;
 
 class UserController extends Controller
 {
@@ -49,9 +52,9 @@ class UserController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
+    public function show()
     {
-        //
+        return view('user.pages.profiles.show');
     }
 
     /**
@@ -62,6 +65,7 @@ class UserController extends Controller
      */
     public function edit()
     {
+        // dd(Auth::user()->hasVerifiedEmail());
         $profile=Auth::guard('web')->user()->profile;
         return view('user.pages.profiles.update');
     }
@@ -73,11 +77,15 @@ class UserController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(UpdateRequest $request)
+    public function update(UpdateProfileRequest $request)
     {
-            $file_name="";
-            $oldImage = public_path('storage/'.Auth::user()->profile->profile_img);
-            File::delete($oldImage);
+            $file_name=isset(Auth::user()->profile->profile_img)?Auth::user()->profile->profile_img:NULL;
+            if(isset(Auth::user()->profile->profile_img)){
+                if(($request->file('profile_img'))){
+                $oldImage = public_path('storage/'.Auth::user()->profile->profile_img);
+                File::delete($oldImage);
+                }
+            }
             if($request->file('profile_img')){
                 $img_uploaded=$request->file('profile_img')->store('public');
                 $file_name=$request->file('profile_img')->hashName();
@@ -96,13 +104,31 @@ class UserController extends Controller
                         'gender'=>$request->gender
                     ]
                 );
+                User::find(Auth::id())->update(['name'=>$request->name]);
                 if($profile_updated){
                     return redirect()->route('home.profile.update')->with('success','Profile Updated Successfully');
                 }else{
                     return redirect()->route('home.profile.update')->with('error','Some error occure! Plz contact with developer.');
                 }
     }
-
+    public function changePassword()
+    {
+        // dd(session());
+        return view('user.pages.profiles.change-password');
+    }
+    public function updatePassword(ChangePassword $request)
+    {
+        
+        // dd($request);
+        $hashed_password=Hash::make($request->password);
+        session()->put('password_hash_web',$hashed_password);
+        // return redirect()->route('home.change_password')->with('success','Password Changed Successfully');
+        // dd(session());
+        $changedPassword=User::find(Auth::id())->update(['password'=>$hashed_password]);
+        if($changedPassword){
+            return redirect()->route('home.change_password')->with('success','Password Changed Successfully');
+        }
+    }
     /**
      * Remove the specified resource from storage.
      *
